DDharmonize_validate_BirthCounts

This script implements a workflow for births counts from vr or census extracting data from Demodata, harmonizing age groups, identifying full series, selecting preferred series, validating totals by age and, eventually including key fields in the function output.

DDharmonize_validate_BirthCounts(
  locid,
  times,
  process = c("census", "vr"),
  return_unique_ref_period = TRUE,
  DataSourceShortName = NULL,
  DataSourceYear = NULL,
  retainKeys = FALSE,
  server = "https://popdiv.dfs.un.org/DemoData/api/"
)

Arguments

locid
location id

times
The period of the data to be extracted. You can extract one year data e.g times = 2020 or a longer period of time e.g times = c(1950, 2020).

process
The process through which the data was obtained from various sources i.e either via census or vital registrations (vr). By default, the function pulls data obtained through both of these processes.

return_unique_ref_period
TRUE

DataSourceShortName
NULL

DataSourceYear
NULL

retainKeys
FALSE

server
"https://popdiv.dfs.un.org/DemoData/api/"

Value

A harmonized dataset containing birth counts

Examples

sweden_df <- DDharmonize_validate_BirthCounts(752,
                                              c(1950,2020),
                                              process = c("census", "vr"),
                                              return_unique_ref_period = TRUE,
                                              DataSourceShortName = NULL,
                                              DataSourceYear = NULL,
                                              retainKeys = FALSE,
                                              server = "https://popdiv.dfs.un.org/DemoData/api/")
#> 
#>  https://popdiv.dfs.un.org/DemoData/api/structureddatarecords?dataProcessIds=2,36&startYear=1950&endYear=2020&indicatorIds=159,170&locIds=752&locAreaTypeIds=2&subGroupIds=2 
#> 
#>  https://popdiv.dfs.un.org/DemoData/api/datacatalogs?dataProcessTypeIds=2&locIds=752&addDefault=false 
#> 
#>  https://popdiv.dfs.un.org/DemoData/api/locations?addDefault=false&includeDependencies=false&includeFormerCountries=false 
#> 
#>  https://popdiv.dfs.un.org/DemoData/api/dataprocesses?addDefault=false 
#> 
#>  https://popdiv.dfs.un.org/DemoData/api/dataprocesstypes?addDefault=false 
#> 
#>  https://popdiv.dfs.un.org/DemoData/api/datasources? 
#> 
#>  https://popdiv.dfs.un.org/DemoData/api/datasources?dataProcessTypeIds=2,36&locIds=752&addDefault=false 
#> [1] 1
#> [1] "752 - Sweden - VR - Births - 1950 - Register - Demographic Yearbook - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] 2
#> [1] "752 - Sweden - VR - Births - 1951 - Register - Demographic Yearbook - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] 3
#> [1] "752 - Sweden - VR - Births - 1952 - Register - Demographic Yearbook - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] 4
#> [1] "752 - Sweden - VR - Births - 1953 - Register - Demographic Yearbook - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] 5
#> [1] "752 - Sweden - VR - Births - 1954 - Register - Demographic Yearbook - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] 6
#> [1] "752 - Sweden - VR - Births - 1955 - Register - Demographic Yearbook - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] 7
#> [1] "752 - Sweden - VR - Births - 1956 - Register - Demographic Yearbook - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] 8
#> [1] "752 - Sweden - VR - Births - 1957 - Register - Demographic Yearbook - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] 9
#> [1] "752 - Sweden - VR - Births - 1958 - Register - Demographic Yearbook - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] 10
#> [1] "752 - Sweden - VR - Births - 1959 - Register - Demographic Yearbook - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] 11
#> [1] "752 - Sweden - VR - Births - 1960 - Register - Demographic Yearbook - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] 12
#> [1] "752 - Sweden - VR - Births - 1961 - Register - Demographic Yearbook - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] 13
#> [1] "752 - Sweden - VR - Births - 1962 - Register - Demographic Yearbook - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] 14
#> [1] "752 - Sweden - VR - Births - 1963 - Register - Demographic Yearbook - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] 15
#> [1] "752 - Sweden - VR - Births - 1964 - Register - Demographic Yearbook - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] 16
#> [1] "752 - Sweden - VR - Births - 1965 - Register - Demographic Yearbook - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] 17
#> [1] "752 - Sweden - VR - Births - 1966 - Register - Demographic Yearbook - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] 18
#> [1] "752 - Sweden - VR - Births - 1967 - Register - Demographic Yearbook - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] 19
#> [1] "752 - Sweden - VR - Births - 1968 - Register - Demographic Yearbook - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] 20
#> [1] "752 - Sweden - VR - Births - 1969 - Register - Demographic Yearbook - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] 21
#> [1] "752 - Sweden - VR - Births - 1970 - Register - Demographic Yearbook - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] 22
#> [1] "752 - Sweden - VR - Births - 1971 - Register - Demographic Yearbook - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] 23
#> [1] "752 - Sweden - VR - Births - 1972 - Register - Demographic Yearbook - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] 24
#> [1] "752 - Sweden - VR - Births - 1973 - Register - Demographic Yearbook - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] 25
#> [1] "752 - Sweden - VR - Births - 1974 - Register - Demographic Yearbook - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] 26
#> [1] "752 - Sweden - VR - Births - 1975 - Register - Demographic Yearbook - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] 27
#> [1] "752 - Sweden - VR - Births - 1976 - Register - Demographic Yearbook - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] 28
#> [1] "752 - Sweden - VR - Births - 1977 - Register - Demographic Yearbook - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] 29
#> [1] "752 - Sweden - VR - Births - 1978 - Register - Demographic Yearbook - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] 30
#> [1] "752 - Sweden - VR - Births - 1979 - Register - Demographic Yearbook - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] 31
#> [1] "752 - Sweden - VR - Births - 1980 - Register - Demographic Yearbook - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] 32
#> [1] "752 - Sweden - VR - Births - 1981 - Register - Demographic Yearbook - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] 33
#> [1] "752 - Sweden - VR - Births - 1982 - Register - Demographic Yearbook - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] 34
#> [1] "752 - Sweden - VR - Births - 1983 - Register - Demographic Yearbook - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] 35
#> [1] "752 - Sweden - VR - Births - 1984 - Register - Demographic Yearbook - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] 36
#> [1] "752 - Sweden - VR - Births - 1985 - Register - Demographic Yearbook - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] 37
#> [1] "752 - Sweden - VR - Births - 1986 - Register - Demographic Yearbook - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] 38
#> [1] "752 - Sweden - VR - Births - 1987 - Register - Demographic Yearbook - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] 39
#> [1] "752 - Sweden - VR - Births - 1988 - Register - Demographic Yearbook - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] 40
#> [1] "752 - Sweden - VR - Births - 1989 - Register - Demographic Yearbook - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] 41
#> [1] "752 - Sweden - VR - Births - 1990 - Register - Demographic Yearbook - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] 42
#> [1] "752 - Sweden - VR - Births - 1991 - Register - Demographic Yearbook - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] 43
#> [1] "752 - Sweden - VR - Births - 1992 - Register - Demographic Yearbook - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] 44
#> [1] "752 - Sweden - VR - Births - 1993 - Register - Demographic Yearbook - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] 45
#> [1] "752 - Sweden - VR - Births - 1994 - Register - Demographic Yearbook - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] 46
#> [1] "752 - Sweden - VR - Births - 1995 - Register - Demographic Yearbook - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] 47
#> [1] "752 - Sweden - VR - Births - 1996 - Register - Demographic Yearbook - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] 48
#> [1] "752 - Sweden - VR - Births - 1997 - Register - Demographic Yearbook - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] 49
#> [1] "752 - Sweden - VR - Births - 1998 - Register - Demographic Yearbook - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] 50
#> [1] "752 - Sweden - VR - Births - 1999 - Register - Demographic Yearbook - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] 51
#> [1] "752 - Sweden - VR - Births - 2001 - Register - Demographic Yearbook - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] 52
#> [1] "752 - Sweden - VR - Births - 2002 - Register - Demographic Yearbook - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] 53
#> [1] "752 - Sweden - VR - Births - 2004 - Register - Demographic Yearbook - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] 54
#> [1] "752 - Sweden - VR - Births - 2005 - Register - Demographic Yearbook - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] 55
#> [1] "752 - Sweden - VR - Births - 2006 - Register - Demographic Yearbook - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] "harmonizing vital counts by 1-year age group"
#> [1] 57
#> [1] "752 - Sweden - VR - Births - 2007 - Register - Eurostat Database - Year of occurrence - Direct - High quality"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] "harmonizing vital counts by 1-year age group"
#> [1] "harmonizing vital counts by 1-year age group"
#> [1] 59
#> [1] "752 - Sweden - VR - Births - 2008 - Register - Eurostat Database - Year of occurrence - Direct - High quality"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] "harmonizing vital counts by 1-year age group"
#> [1] "harmonizing vital counts by 1-year age group"
#> [1] 61
#> [1] "752 - Sweden - VR - Births - 2009 - Register - Eurostat Database - Year of occurrence - Direct - High quality"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] "harmonizing vital counts by 1-year age group"
#> [1] "harmonizing vital counts by 1-year age group"
#> [1] 63
#> [1] "752 - Sweden - VR - Births - 2010 - Register - Eurostat Database - Year of occurrence - Direct - High quality"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] "harmonizing vital counts by 1-year age group"
#> [1] "harmonizing vital counts by 1-year age group"
#> [1] 65
#> [1] "752 - Sweden - VR - Births - 2011 - Register - Eurostat Database - Year of occurrence - Direct - High quality"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] "harmonizing vital counts by 1-year age group"
#> [1] "harmonizing vital counts by 1-year age group"
#> [1] 67
#> [1] "752 - Sweden - VR - Births - 2012 - Register - Eurostat Database - Year of occurrence - Direct - High quality"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] "harmonizing vital counts by 1-year age group"
#> [1] 68
#> [1] "752 - Sweden - VR - Births - 2013 - Register - Demographic Yearbook - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] "harmonizing vital counts by 1-year age group"
#> [1] 69
#> [1] "752 - Sweden - VR - Births - 2013 - Register - Eurostat Database - Year of occurrence - Direct - High quality"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] "harmonizing vital counts by 1-year age group"
#> [1] 70
#> [1] "752 - Sweden - VR - Births - 2013 - Register - Statistics Sweden website - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] 71
#> [1] "752 - Sweden - VR - Births - 2014 - Register - Demographic Yearbook - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] "harmonizing vital counts by 1-year age group"
#> [1] 72
#> [1] "752 - Sweden - VR - Births - 2014 - Register - Eurostat Database - Year of occurrence - Direct - High quality"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] "harmonizing vital counts by 1-year age group"
#> [1] 73
#> [1] "752 - Sweden - VR - Births - 2014 - Register - Statistics Sweden website - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] 74
#> [1] "752 - Sweden - VR - Births - 2015 - Register - Demographic Yearbook - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] "harmonizing vital counts by 1-year age group"
#> [1] 75
#> [1] "752 - Sweden - VR - Births - 2015 - Register - Eurostat Database - Year of occurrence - Direct - High quality"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] "harmonizing vital counts by 1-year age group"
#> [1] 76
#> [1] "752 - Sweden - VR - Births - 2016 - Register - Demographic Yearbook - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] "harmonizing vital counts by 1-year age group"
#> [1] 77
#> [1] "752 - Sweden - VR - Births - 2016 - Register - Eurostat Database - Year of occurrence - Direct - High quality"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] "harmonizing vital counts by 1-year age group"
#> [1] 78
#> [1] "752 - Sweden - VR - Births - 2017 - Register - Demographic Yearbook - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] "harmonizing vital counts by 1-year age group"
#> [1] 79
#> [1] "752 - Sweden - VR - Births - 2017 - Register - Eurostat Database - Year of occurrence - Direct - High quality"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] "harmonizing vital counts by 1-year age group"
#> [1] 80
#> [1] "752 - Sweden - VR - Births - 2018 - Register - Demographic Yearbook - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] "harmonizing vital counts by 1-year age group"
#> [1] 81
#> [1] "752 - Sweden - VR - Births - 2018 - Register - Eurostat Database - Year of occurrence - Direct - High quality"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] "harmonizing vital counts by 1-year age group"
#> [1] 82
#> [1] "752 - Sweden - VR - Births - 2019 - Register - Demographic Yearbook - Year of occurrence - Direct - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] "harmonizing vital counts by 1-year age group"
#> [1] 83
#> [1] "752 - Sweden - VR - Births - 2019 - Register - Eurostat Database - Year of occurrence - Direct - High quality"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] "harmonizing vital counts by 1-year age group"
#> 
#>  Location ID:  752 
#>  Location Name:  Sweden