DDharmonize_validate_DeathCounts

This script implements a workflow for death counts from vr or census extracting data from Demodata, harmonizing age groups, identifying full series, selecting preferred series, validating totals by age and by sex and eventually, including key fields in the function output.

DDharmonize_validate_DeathCounts(
  locid,
  times,
  process = c("census", "vr"),
  return_unique_ref_period = TRUE,
  DataSourceShortName = NULL,
  DataSourceYear = NULL,
  retainKeys = FALSE,
  server = "https://popdiv.dfs.un.org/DemoData/api/"
)

Arguments

locid
location id

times
1950, 2050

process
census or vr

return_unique_ref_period
TRUE

DataSourceShortName
NULL

DataSourceYear
NULL

retainKeys
FALSE

server
"https://popdiv.dfs.un.org/DemoData/api/"

Value

A harmonized dataset containing death counts

Examples

kenya_df <- DDharmonize_validate_DeathCounts(404,
                                              c(1950,2020),
                                              process = c("census", "vr"),
                                              return_unique_ref_period = TRUE,
                                              DataSourceShortName = NULL,
                                              DataSourceYear = NULL,
                                              retainKeys = FALSE,
                                              server = "https://popdiv.dfs.un.org/DemoData/api/")

#> 
#>  https://popdiv.dfs.un.org/DemoData/api/structureddatarecords?dataProcessIds=2,36&startYear=1950&endYear=2020&indicatorIds=194,195,188&locIds=404&locAreaTypeIds=2&subGroupIds=2 
#> 
#>  https://popdiv.dfs.un.org/DemoData/api/datacatalogs?dataProcessTypeIds=2&locIds=404&addDefault=false 
#> 
#>  https://popdiv.dfs.un.org/DemoData/api/locations?addDefault=false&includeDependencies=false&includeFormerCountries=false 
#> 
#>  https://popdiv.dfs.un.org/DemoData/api/dataprocesses?addDefault=false 
#> 
#>  https://popdiv.dfs.un.org/DemoData/api/dataprocesstypes?addDefault=false 
#> 
#>  https://popdiv.dfs.un.org/DemoData/api/datasources? 
#> 
#>  https://popdiv.dfs.un.org/DemoData/api/datasources?dataProcessTypeIds=2,36&locIds=404&addDefault=false 
#> [1] 1
#> [1] "404 - Kenya - Census - Deaths - 2009 - Census - Kenya. Population and Housing Census 2009 - De-facto - Household deaths - Fair"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] "harmonizing vital counts by 1-year age group"
#> [1] 2
#> [1] "404 - Kenya - Census - Deaths - 2019 - Census - Demographic Yearbook - De-facto - Direct - Unknown"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] "harmonizing vital counts by 1-year age group"
#> [1] 3
#> [1] "404 - Kenya - Census - Deaths - 2019 - Census - Demographic Yearbook - De-facto - Household deaths - Unknown"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] "harmonizing vital counts by 1-year age group"
#> [1] 4
#> [1] "404 - Kenya - VR - Deaths - 1960 - Register - Demographic Yearbook - Year of registration - Direct - Low"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] "harmonizing vital counts by 1-year age group"
#> [1] 5
#> [1] "404 - Kenya - VR - Deaths - 1961 - Register - Demographic Yearbook - Year of registration - Direct - Low"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] "harmonizing vital counts by 1-year age group"
#> [1] 6
#> [1] "404 - Kenya - VR - Deaths - 1962 - Register - Demographic Yearbook - Year of registration - Direct - Low"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] "harmonizing vital counts by 1-year age group"
#> [1] 7
#> [1] "404 - Kenya - VR - Deaths - 1963 - Register - Demographic Yearbook - Year of registration - Direct - Low"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] "harmonizing vital counts by 1-year age group"
#> [1] 8
#> [1] "404 - Kenya - VR - Deaths - 1968 - Register - Demographic Yearbook - Year of registration - Direct - Low"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] "harmonizing vital counts by 1-year age group"
#> [1] 9
#> [1] "404 - Kenya - VR - Deaths - 1968 - Register - Mortality in developing countries - Unknown - Unknown/Not mentioned - Low"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] "harmonizing vital counts by 1-year age group"
#> [1] 10
#> [1] "404 - Kenya - VR - Deaths - 1969 - Register - Demographic Yearbook - Year of registration - Direct - Low"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] "harmonizing vital counts by 1-year age group"
#> [1] 11
#> [1] "404 - Kenya - VR - Deaths - 1969 - Register - Mortality in developing countries - Unknown - Unknown/Not mentioned - Low"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] "harmonizing vital counts by 1-year age group"
#> [1] 12
#> [1] "404 - Kenya - VR - Deaths - 1970 - Register - Demographic Yearbook - Year of registration - Direct - Low"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] "harmonizing vital counts by 1-year age group"
#> [1] 13
#> [1] "404 - Kenya - VR - Deaths - 1970 - Register - Mortality in developing countries - Unknown - Unknown/Not mentioned - Low"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] "harmonizing vital counts by 1-year age group"
#> [1] 14
#> [1] "404 - Kenya - VR - Deaths - 1971 - Register - Demographic Yearbook - Year of registration - Direct - Low"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] "harmonizing vital counts by 1-year age group"
#> [1] 15
#> [1] "404 - Kenya - VR - Deaths - 1972 - Register - Demographic Yearbook - Year of registration - Direct - Low"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] "harmonizing vital counts by 1-year age group"
#> [1] 16
#> [1] "404 - Kenya - VR - Deaths - 1973 - Register - Demographic Yearbook - Year of registration - Direct - Low"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] "harmonizing vital counts by 1-year age group"
#> [1] 17
#> [1] "404 - Kenya - VR - Deaths - 1973 - Register - Mortality in developing countries - Unknown - Unknown/Not mentioned - Low"
#> [1] "harmonizing vital counts by 5-year age group"
#> [1] "harmonizing vital counts by 1-year age group"
#> [1] "harmonizing vital counts by 1-year age group"
#> [1] "harmonizing vital counts by 1-year age group"
#> [1] "harmonizing vital counts by 1-year age group"
#> [1] "harmonizing vital counts by 1-year age group"
#> [1] "harmonizing vital counts by 1-year age group"
#> [1] "harmonizing vital counts by 1-year age group"
#> [1] "harmonizing vital counts by 1-year age group"
#> [1] "harmonizing vital counts by 1-year age group"
#> 
#>  Location ID:  404 
#>  Location Name:  Kenya